pyproject.toml
[project]
name = "bone-dot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = ["fake-bpy-module>=20250414"]

[tool.pyright]
typeCheckingMode = "off"
